html/semantics/forms/form-submission-0/form-double-submit-two-targets.html

<!DOCTYPE html>
<meta charset="utf-8">
<link rel="author" title="Joey Arhar" href="mailto:jarhar@chromium.org">
<link rel="help" href="https://html.spec.whatwg.org/C/#plan-to-navigate">

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<form id=myform name=myform action="/formaction.html"></form>
<iframe id=frame1 name=target1></iframe>
<iframe id=frame2 name=target2></iframe>

<script>

promise_test(async () => {
  const frame1LoadPromise = new Promise(resolve => frame1.onload = resolve);
  const frame2LoadPromise = new Promise(resolve => frame2.onload = resolve);

  myform.target = 'target1';
  myform.submit();
  myform.target = 'target2';
  myform.submit();

  await Promise.all([frame1LoadPromise, frame2LoadPromise]);

  assert_equals(frame1.contentDocument.location.pathname, '/formaction.html');
  assert_equals(frame2.contentDocument.location.pathname, '/formaction.html');

}, 'Verifies that one form used to target multiple frames in succession navigates both of them.');

</script>
